move method to node-api to address review comments
