+++ node/src/main/kotlin/net/corda/node/internal/NodeStartup.kt
@@ -29,6 +29,7 @@ import net.corda.node.services.config.NodeConfiguration
 import net.corda.node.services.config.shouldStartLocalShell
 import net.corda.node.services.config.shouldStartSSHDaemon
 import net.corda.node.utilities.registration.NodeRegistrationException
+import net.corda.nodeapi.internal.JVMAgentUtilities
 import net.corda.nodeapi.internal.addShutdownHook
 import net.corda.nodeapi.internal.persistence.CouldNotCreateDataSourceException
 import net.corda.nodeapi.internal.persistence.DatabaseIncompatibleException
@@ -153,23 +154,7 @@ open class NodeStartup : NodeStartupLogging {
         const val LOGS_CAN_BE_FOUND_IN_STRING = "Logs can be found in"
         const val ERROR_CODE_RESOURCE_LOCATION = "error-codes"
 
-        @VisibleForTesting
-        @Suppress("NestedBlockDepth")
-        fun parseDebugPort(args : Iterable<String>) : Short? {
-            val debugArgumentPrefix = "-agentlib:jdwp="
-            for(arg in args) {
-                if(arg.startsWith(debugArgumentPrefix)) {
-                    for(keyValuePair in arg.substring(debugArgumentPrefix.length + 1).split(",")) {
-                        val equal = keyValuePair.indexOf('=')
-                        if(equal >= 0 && keyValuePair.startsWith("address")) {
-                            val portBegin = (keyValuePair.lastIndexOf(':').takeUnless { it < 0 } ?: equal) + 1
-                            return keyValuePair.substring(portBegin).toShort()
-                        }
-                    }
-                }
-            }
-            return null
-        }
+
     }
 
     lateinit var cmdLineOptions: SharedNodeCmdLineOptions
@@ -286,7 +271,7 @@ open class NodeStartup : NodeStartupLogging {
         logger.info("VM ${info.vmName} ${info.vmVendor} ${info.vmVersion}")
         logger.info("Machine: ${lookupMachineNameAndMaybeWarn()}")
         logger.info("Working Directory: ${cmdLineOptions.baseDirectory}")
-        parseDebugPort(info.inputArguments) ?.let {
+        JVMAgentUtilities.parseDebugPort(info.inputArguments) ?.let {
             logger.info("Debug port: $it")
         }
